Added dto.parseRef(): Reduce the dependency to keytype.

+++ icsf/identity/icsf-acl/src/main/java/com/bee32/icsf/principal/IPrincipal.java
@@ -12,7 +12,7 @@ import com.bee32.plover.orm.entity.IEntity;
  * 在涉及安全的操作中，代表行为主体的身份抽象。
  */
 public interface IPrincipal
-        extends Principal, IEntity<Long> {
+        extends Principal, IEntity<Integer> {
 
     /**
      * 主体在该主体所属域中的唯一名称。
