+++ icsf/identity/icsf-acl/src/main/java/com/bee32/icsf/principal/dao/AbstractPrincipalDao.java
@@ -6,7 +6,7 @@ import com.bee32.icsf.principal.Principal;
 import com.bee32.plover.orm.entity.EntityDao;
 
 public class AbstractPrincipalDao<E extends Principal>
-        extends EntityDao<E, Long> {
+        extends EntityDao<E, Integer> {
 
     public E getByName(String name) {
         E user = (E) getSession().createCriteria(getEntityType())//
