+++ icsf/identity/icsf-acl/src/main/java/com/bee32/icsf/principal/dto/AbstractPrincipalDto.java
@@ -7,10 +7,9 @@ import com.bee32.icsf.principal.Principal;
 import com.bee32.plover.arch.util.Mask32;
 import com.bee32.plover.arch.util.TextMap;
 import com.bee32.plover.orm.util.EntityDto;
-import com.bee32.plover.orm.util.IEntityMarshalContext;
 
 public class AbstractPrincipalDto<E extends Principal>
-        extends EntityDto<E, Long> {
+        extends EntityDto<E, Integer> {
 
     private static final long serialVersionUID = 1L;
